moved props closer to components, which saved a lot of rerenders
