+++ src/components/Legend/styles.js
