+++ src/components/Tile/styles.js
