+++ src/sections/Details/styles.js
